Get rid of auth.service and extra token methods
