+++ template/apps/api/src/resources/account/endpoints/signOut.ts
@@ -1,9 +1,17 @@
 import { pub } from 'procedures';
 import { z } from 'zod';
 
-import { authService } from 'services';
+import { tokensService } from 'db';
+import { TokenType } from 'resources/tokens/tokens.schema';
+
+import { cookieUtil } from 'utils';
 
 export default pub.output(z.object({})).handler(async ({ context }) => {
-  await authService.unsetUserAccessToken({ ctx: context });
+  if (context.user) {
+    await tokensService.deleteMany({ userId: context.user._id, type: TokenType.ACCESS });
+  }
+
+  await cookieUtil.unsetTokens({ ctx: context });
+
   return {};
 });
