+++ template/apps/api/src/resources/account/endpoints/verifyEmailToken.ts
@@ -3,12 +3,13 @@ import { z } from 'zod';
 
 import { TokenType } from 'resources/tokens/tokens.schema';
 import validateToken from 'resources/tokens/methods/validateToken';
-import invalidateUserTokens from 'resources/tokens/methods/invalidateUserTokens';
+
 import { publicSchema } from 'resources/users/users.schema';
 import getPublic from 'resources/users/methods/getPublic';
-import { usersService } from 'db';
+import { usersService, tokensService } from 'db';
 
-import { authService, emailService } from 'services';
+import setAccessToken from 'resources/tokens/methods/setAccessToken';
+import { emailService } from 'services';
 
 import config from 'config';
 
@@ -35,10 +36,10 @@ export default pub
       throw new ClientError({ token: 'Token is invalid or expired' });
     }
 
-    await invalidateUserTokens(user._id, TokenType.EMAIL_VERIFICATION);
+    await tokensService.deleteMany({ userId: user._id, type: TokenType.EMAIL_VERIFICATION });
     await usersService.updateOne({ _id: user._id }, () => ({ isEmailVerified: true }));
 
-    const accessToken = await authService.setAccessToken({ ctx: context, userId: user._id });
+    const accessToken = await setAccessToken({ ctx: context, userId: user._id });
 
     await emailService.sendTemplate<typeof Template.SIGN_UP_WELCOME>({
       to: user.email,
