+++ template/apps/api/src/services/index.ts
@@ -1,8 +1,7 @@
 import analyticsService from './analytics/analytics.service';
-import * as authService from './auth/auth.service';
 import cloudStorageService from './cloud-storage/cloud-storage.service';
 import emailService from './email/email.service';
 import * as googleService from './google/google.service';
 import socketService from './socket/socket.service';
 
-export { analyticsService, authService, cloudStorageService, emailService, googleService, socketService };
+export { analyticsService, cloudStorageService, emailService, googleService, socketService };
